backend/requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
google-generativeai
PyPDF2
python-docx
easyocr
torch
torchvision
torchaudio